bundle update
bundle install
...
Your bundle is locked to codecov (0.1.14) from rubygems repository https://rubygems.org/ or installed locally, but that version can
no longer be found in that source. That means the author of codecov (0.1.14) has removed it. You'll need to update your bundle to a
version other than codecov (0.1.14) that hasn't been removed in order to install.